datatype changed

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,9 +8,9 @@
 
 int main(void)
 {
-unsigned long long int start = 0;
-unsigned long long int nextdigit = 1;
-unsigned long long int sum = 0;
+unsigned long int start = 0;
+unsigned long int nextdigit = 1;
+unsigned long int sum = 0;
 unsigned long number = 0;
 
 while (number < 101)
